backend/services/subscription-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
boto3>=1.28.0
pyjwt>=2.8.0
cryptography>=41.0.0
stripe>=7.0.0

